dataset_builder/requirements.txt
numpy
Pillow
openai
python-dotenv
opencv-python
scikit-image
tqdm